fixup! remove envelope
